tests/wpt/tests/html/webappapis/dynamic-markup-insertion/document-write/044.html

<!doctype html>
<title>document.write</title>
<script src="/resources/testharness.js"></script><script src="/resources/testharnessreport.js"></script>
<script>
var t = async_test();
t.step(function() {
  document.write("<body><textarea><span>Filler</span></textarea>");
});
</script>
<script>
t.step(function() {
  assert_equals(document.body.childNodes[0].localName, "textarea");
  assert_equals(document.body.childNodes[0].textContent, "<span>Filler</span>");
})
t.done();
</script>
<div id="log"></div>
